### 2021-04-17 #3188 @whedon
[PRE REVIEW]: The stantargets R package: a workflow framework for efficient reproducible Stan-powered Bayesian data analysis pipelines
**Submitting author:** @wlandau (<a href="http://orcid.org/0000-0003-1878-3253">William Landau</a>)
**Repository:** <a href="https://github.com/ropensci/stantargets" target ="_blank">https://github.com/ropensci/stantargets</a>
**Version:** 0.0.0.9003
