### 2021-04-17 #3188 @kyleniemeyer
Hi @wlandau, since this submission was already reviewed and accepted by rOpenSci, the review at JOSS will be fast-tracked.

rOpenSci review issue: https://github.com/ropensci/software-review/issues/430
