### 2021-04-17 #3188 @kyleniemeyer
@wlandau however, it does look like your paper is missing the required statement of need section, one of the [required elements](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) of JOSS papers. Can you add this, then we will proceed?
