### 2021-04-17 #3188 @wlandau
Thanks @kyleniemeyer, https://github.com/ropensci/stantargets/commit/bd5e306d05c60274b45eb584e83ef76ea3158716 adds a statement of need section.
