### 2021-04-19 #3193 @kyleniemeyer
Hi @wlandau, I just noticed a few minor issues with some references, it looks like some need additional information or different formatting. In particular, the Cndstanr and Stantargets citations should both include a URL or DOI (if one exists), and both should have years (e.g., for the latest version, if you aren't citing a particular version).
