### 2021-04-20 #3193 @wlandau
Thanks @kyleniemeyer, should be fixed in https://github.com/ropensci/stantargets/commit/1f9b58f795189e5fba02596b5e772092b0c7dae1.
